refactor(test): RaceResultTest 테스트 클래스 추가
