+++ src/test/kotlin/race/RaceTest.kt
@@ -26,20 +26,4 @@ class RaceTest {
             { assertThat(records[2][1].position).isEqualTo(0) }
         )
     }
-
-    @Test
-    fun `우승자는 car1과 car3`() {
-        val trueCondition = FixedMoveCondition(true)
-        val car1 = Car("car1", trueCondition)
-
-        val falseCondition = FixedMoveCondition(false)
-        val car2 = Car("car2", falseCondition)
-
-        val car3 = Car("car3", trueCondition)
-
-        val raceResult = Race().run(listOf(car1, car2, car3), 2)
-        val winners = raceResult.findWinners()
-        assertThat(winners[0].name).isEqualTo("car1")
-        assertThat(winners[1].name).isEqualTo("car3")
-    }
 }
